Added exception

+++ weather-service/src/main/java/com/fabrick/interview/weather/service/WeatherService.java
@@ -5,8 +5,6 @@ import com.fabrick.interview.weather.model.Station;
 import reactor.core.publisher.Flux;
 
 public interface WeatherService {
-
-    Flux<Station> findStationsCloseToAirport(String airportId, float closestBy);
-    Flux<Airport> findAirportsCloseToStation(String stationId, float closestBy);
-
+    Flux<Station> findStationsCloseToAirport(String airportId, double closestBy);
+    Flux<Airport> findAirportsCloseToStation(String stationId, double closestBy);
 }
